requirements.txt
pyyaml
selenium
apscheduler
uiautomator2
pillow
weditor